+++ test/sql/type.sql
@@ -7,10 +7,6 @@
 CREATE TABLE h3_test_type (hex h3index PRIMARY KEY);
 INSERT INTO h3_test_type (hex) SELECT * from h3_get_res_0_indexes();
 
--- Type conversions
-SELECT h3_string_to_h3(h3_to_string(:hexagon)) = :hexagon;
-SELECT h3_to_string(h3_string_to_h3(:string))::text = :string::text;
-
 --
 -- TEST operators
 --